bugfix: convert metainfo from http style to rpc style
In previous codes, metainfo does not have a uniform style, and it
cannot work properly between different types of service.

We agree to use following rules:

- For RPC, use uppercase letters and `_` as separators when
  transmitting
- For HTTP, use lowercase letters and `-` as separators when
  transmitting
- For both protocols, use RPC style (uppercase with `_`) for storing
  keys of metainfo.

To implement this more elegantly, we defined a trait `Converter` and
use it for converting the key names between transmitting format and
storing format.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>